backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pymongo==4.8.0
python-dotenv==1.0.1
pydantic==2.8.2
aiortc==1.9.0
numpy==2.1.1
soundfile==0.12.1
faster-whisper==1.0.3
google-generativeai==0.7.2
requests==2.32.3
pyttsx3==2.90
